# [sonic-cfggen] remove lazy_re workaround due to many optimizations done