… src/sonic-config-engine/lazy_re.py